Move util functions from CommonUtils to RecommendationUtils
