+++ src/main/java/com/autotune/common/utils/CommonUtils.java
@@ -335,164 +335,4 @@ public class CommonUtils {
         return datasource == null || datasource.getAuthenticationConfig() == null ||
                 datasource.getAuthenticationConfig().toString().isEmpty();
     }
-
-    public static void markAcceleratorDeviceStatusToContainer (ContainerData containerData,
-                                                               String maxDateQuery,
-                                                               String namespace,
-                                                               String workload,
-                                                               String workload_type,
-                                                               DataSourceInfo dataSourceInfo,
-                                                               Map<String, Terms> termsMap,
-                                                               Double measurementDurationMinutesInDouble,
-                                                               String gpuDetectionQuery)
-            throws IOException, NoSuchAlgorithmException, KeyStoreException,
-            KeyManagementException, ParseException, FetchMetricsError {
-
-        SimpleDateFormat sdf = new SimpleDateFormat(KruizeConstants.DateFormats.STANDARD_JSON_DATE_FORMAT, Locale.ROOT);
-        String containerName = containerData.getContainer_name();
-        String queryToEncode = null;
-        long interval_end_time_epoc = 0;
-        long interval_start_time_epoc = 0;
-
-        LOGGER.debug("maxDateQuery: {}", maxDateQuery);
-        queryToEncode =  maxDateQuery
-                .replace(AnalyzerConstants.NAMESPACE_VARIABLE, namespace)
-                .replace(AnalyzerConstants.CONTAINER_VARIABLE, containerName)
-                .replace(AnalyzerConstants.WORKLOAD_VARIABLE, workload)
-                .replace(AnalyzerConstants.WORKLOAD_TYPE_VARIABLE, workload_type);
-
-        String dateMetricsUrl = String.format(KruizeConstants.DataSourceConstants.DATE_ENDPOINT_WITH_QUERY,
-                dataSourceInfo.getUrl(),
-                URLEncoder.encode(queryToEncode, CHARACTER_ENCODING)
-        );
-
-        LOGGER.debug(dateMetricsUrl);
-        GenericRestApiClient client = new GenericRestApiClient(dataSourceInfo);
-        client.setBaseURL(dateMetricsUrl);
-        JSONObject genericJsonObject = client.fetchMetricsJson(KruizeConstants.APIMessages.GET, "");
-        JsonObject jsonObject = new Gson().fromJson(genericJsonObject.toString(), JsonObject.class);
-        JsonArray resultArray = jsonObject.getAsJsonObject(KruizeConstants.JSONKeys.DATA).getAsJsonArray(KruizeConstants.DataSourceConstants.DataSourceQueryJSONKeys.RESULT);
-
-        if (null == resultArray || resultArray.isEmpty()) {
-            // Need to alert that container max duration is not detected
-            // Ignoring it here, as we take care of it at generate recommendations
-            return;
-        }
-
-        resultArray = resultArray.get(0)
-                .getAsJsonObject().getAsJsonArray(KruizeConstants.DataSourceConstants.DataSourceQueryJSONKeys.VALUE);
-        long epochTime = resultArray.get(0).getAsLong();
-        String timestamp = sdf.format(new Date(epochTime * KruizeConstants.TimeConv.NO_OF_MSECS_IN_SEC));
-        Date date = sdf.parse(timestamp);
-        Timestamp dateTS = new Timestamp(date.getTime());
-        interval_end_time_epoc = dateTS.getTime() / KruizeConstants.TimeConv.NO_OF_MSECS_IN_SEC
-                - ((long) dateTS.getTimezoneOffset() * KruizeConstants.TimeConv.NO_OF_SECONDS_PER_MINUTE);
-        int maxDay = Terms.getMaxDays(termsMap);
-        LOGGER.debug(KruizeConstants.APIMessages.MAX_DAY, maxDay);
-        Timestamp startDateTS = Timestamp.valueOf(Objects.requireNonNull(dateTS).toLocalDateTime().minusDays(maxDay));
-        interval_start_time_epoc = startDateTS.getTime() / KruizeConstants.TimeConv.NO_OF_MSECS_IN_SEC
-                - ((long) startDateTS.getTimezoneOffset() * KruizeConstants.TimeConv.NO_OF_MSECS_IN_SEC);
-
-        gpuDetectionQuery = gpuDetectionQuery.replace(AnalyzerConstants.NAMESPACE_VARIABLE, namespace)
-                .replace(AnalyzerConstants.CONTAINER_VARIABLE, containerName)
-                .replace(AnalyzerConstants.MEASUREMENT_DURATION_IN_MIN_VARAIBLE, Integer.toString(measurementDurationMinutesInDouble.intValue()))
-                .replace(AnalyzerConstants.WORKLOAD_VARIABLE, workload)
-                .replace(AnalyzerConstants.WORKLOAD_TYPE_VARIABLE, workload_type);
-
-        String podMetricsUrl;
-        try {
-            podMetricsUrl = String.format(KruizeConstants.DataSourceConstants.DATASOURCE_ENDPOINT_WITH_QUERY,
-                    dataSourceInfo.getUrl(),
-                    URLEncoder.encode(gpuDetectionQuery, CHARACTER_ENCODING),
-                    interval_start_time_epoc,
-                    interval_end_time_epoc,
-                    measurementDurationMinutesInDouble.intValue() * KruizeConstants.TimeConv.NO_OF_SECONDS_PER_MINUTE);
-            LOGGER.debug(podMetricsUrl);
-            client.setBaseURL(podMetricsUrl);
-            genericJsonObject = client.fetchMetricsJson(KruizeConstants.APIMessages.GET, "");
-
-            jsonObject = new Gson().fromJson(genericJsonObject.toString(), JsonObject.class);
-            resultArray = jsonObject.getAsJsonObject(KruizeConstants.JSONKeys.DATA).getAsJsonArray(KruizeConstants.DataSourceConstants.DataSourceQueryJSONKeys.RESULT);
-
-            if (null != resultArray && !resultArray.isEmpty()) {
-                for (JsonElement result : resultArray) {
-                    JsonObject resultObject = result.getAsJsonObject();
-                    JsonArray valuesArray = resultObject.getAsJsonArray(KruizeConstants.DataSourceConstants
-                            .DataSourceQueryJSONKeys.VALUES);
-
-                    for (JsonElement element : valuesArray) {
-                        JsonArray valueArray = element.getAsJsonArray();
-                        double value = valueArray.get(1).getAsDouble();
-                        // TODO: Check for non-zero values to mark as GPU workload
-                        break;
-                    }
-
-                    JsonObject metricObject = resultObject.getAsJsonObject(KruizeConstants.JSONKeys.METRIC);
-                    String modelName = metricObject.get(KruizeConstants.JSONKeys.MODEL_NAME).getAsString();
-                    if (null == modelName)
-                        continue;
-
-                    boolean isSupportedMig = checkIfModelIsKruizeSupportedMIG(modelName);
-                    if (isSupportedMig) {
-                        AcceleratorDeviceData acceleratorDeviceData = new AcceleratorDeviceData(metricObject.get(KruizeConstants.JSONKeys.MODEL_NAME).getAsString(),
-                                metricObject.get(KruizeConstants.JSONKeys.HOSTNAME).getAsString(),
-                                metricObject.get(KruizeConstants.JSONKeys.UUID).getAsString(),
-                                metricObject.get(KruizeConstants.JSONKeys.DEVICE).getAsString(),
-                                isSupportedMig);
-
-
-                        if (null == containerData.getContainerDeviceList()) {
-                            ContainerDeviceList containerDeviceList = new ContainerDeviceList();
-                            containerData.setContainerDeviceList(containerDeviceList);
-                        }
-                        containerData.getContainerDeviceList().addDevice(AnalyzerConstants.DeviceType.ACCELERATOR, acceleratorDeviceData);
-                        // TODO: Currently we consider only the first mig supported GPU
-                        return;
-                    }
-                }
-            }
-        } catch (IOException | NoSuchAlgorithmException | KeyStoreException | KeyManagementException |
-                 JsonSyntaxException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public static boolean checkIfModelIsKruizeSupportedMIG(String modelName) {
-        if (null == modelName || modelName.isEmpty())
-            return false;
-
-        modelName = modelName.toUpperCase();
-
-        boolean A100_CHECK = (modelName.contains("A100") &&
-                (modelName.contains("40GB") || modelName.contains("80GB")));
-
-        boolean H100_CHECK = false;
-
-        if (!A100_CHECK) {
-            H100_CHECK = (modelName.contains("H100") && modelName.contains("80GB"));
-        }
-
-        return A100_CHECK || H100_CHECK;
-    }
-
-    public static Timestamp getNearestTimestamp(HashMap<Timestamp, IntervalResults> containerDataResults, Timestamp targetTime, int minutesRange) {
-        long rangeInMillis = (long) minutesRange * 60 * 1000;
-        long targetTimeMillis = targetTime.getTime();
-
-        Timestamp nearestTimestamp = null;
-        long nearestDistance = Long.MAX_VALUE;
-
-        for (Map.Entry<Timestamp, IntervalResults> entry : containerDataResults.entrySet()) {
-            Timestamp currentTimestamp = entry.getKey();
-            long currentTimeMillis = currentTimestamp.getTime();
-            long distance = Math.abs(targetTimeMillis - currentTimeMillis);
-
-            if (distance <= rangeInMillis && distance < nearestDistance) {
-                nearestDistance = distance;
-                nearestTimestamp = currentTimestamp;
-            }
-        }
-
-        return nearestTimestamp;
-    }
 }
